Move role capability instance management
In previously, NuguClient had charge of capability instance
management, so, when deinitializing NuguClient, it removed
that instances.

As moving the role of managing capability instance to Application,
it add deinitialize method in capability interface for clearing
relevant resources which is needed to do before removing instance.

When deinitializing NuguClient, that added method is called, so
the Application may remove capability instance freely.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>