Open Base Capability publicly
The Capability class is base class of all Capability Agents
in NUGU SDK and it was hidden from external.

Because, that class inherits ICapabilityInterface and
implements basic functions like handling directive,
sending event, it needs to open publicly which is
possible to use this class for composing add-on capability agent.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>