Implements capability suspend/restore
It implements suspend, restore methods in capability agents
which are required.

In ASRAgent, it implements suspend() for stopRecognition.
In TTSAgent, it implements suspend() for stopTTS.
In AudioPlayerAgent, it implements suspend() for pause/stop
and restore() for resume.

Also, it apply suspend policy which is possible to select
pause or resume action in each capability agent.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>