add method about EpdSilenceInterval
It add the method which is possible to get
epd silence interval (frankly, it's epd pause length).

Because, it's related to ASR, that method is added to ASRAgent.

Be careful, not to use that method before NuguClient initialized.
The reason is at that time, the required instance(SpeechRecognizer)
is not created, we can't retrieve correct value.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>