network: set v2 api to default
Change the default device-gateway-server API v1 to v2.

Below is the status of APIs:

    /v1/event + /v1/event-attachment : deprecated
    /v1/events                       : removed
    /v2/events                       : default

The /v1/events was removed because it was never used by default before.

If you want to use /v1/event, it is possible by setting the environment
variable below. However, it is not recommended because it will be
deleted in the future.

    export NUGU_NETWORK_USE_V1=1

Signed-off-by: Inho Oh <inho.oh@sk.com>