base: introduce profiling feature
The profiling module provides notification and performance measurement
for each condition.

When 'nugu_prof_mark()' or 'nugu_prof_mark_data()' is called, the
timestamp and additional data are stored in the internal cache, and
events can be received through callbacks.

Developers can fetch and analyze the profiled data at the moment they
want through a callback or 'nugu_prof_get_last_data()' API.

In addition, it is possible to simply output the profiling result of
a specific range through the 'nugu_prof_dump()' API.

e.g.

    --------------------------
    Profiling: 0(Created) ~ 38(END)
    [ 0]          Created: <date.msec>   0 msec
    [ 1]      Plugin init: <date.msec>  11 msec [ 0 ~  1]:  11 msec
    [ 2]      Plugin done: <date.msec> 114 msec [ 1 ~  2]: 102 msec
    [ 3]      Initialized: <date.msec> 120 msec [ 0 ~  3]: 120 msec
    --------------------------

Signed-off-by: Inho Oh <inho.oh@sk.com>