base: add application log module
When there is no log system in the target, add a module type so that
the log function provided by the SDK can be used in the application.

Usage in user-app.c:

    #define NUGU_LOG_MODULE NUGU_LOG_MODULE_APPLICATION
    #include <base/nugu_log.h>

    nugu_dbg("checkpoint!");

The application log is enabled by default. And it can be set through
environment variables below.

Disable application log:

    export NUGU_LOG_MODULE=preset_sdk

Enable application log:

    export NUGU_LOG_MODULE=preset_sdk,application
    or
    export NUGU_LOG_MODULE=preset_default

Signed-off-by: Inho Oh <inho.oh@sk.com>