tts: set size limit for the 1st attachment
There is a variation in performance because the size of the first of
the attachment data coming in the TTS response is not uniform.

This patch attempts to ensure constant performance for the first data
output by setting limits (4200 bytes) to the smallest of the first data
from multiple scenario cases.

Signed-off-by: Inho Oh <inho.oh@sk.com>