build: fix ASAN default build option
The default value of the ASAN option is ON only when the build type is
set to Debug in the x86 environment, and the default is OFF otherwise.

However, there was a bug that the `ENABLE_ASAN` option to manually set
the ASAN value did not work in the above environment.

This patch fixes the above bug by presetting environment-specific defaults.

Below is the ASAN status according to the build option and environment.

ASAN default off:

    cmake -DPACKAGING=1               # Debian packaging
    cmake -DCMAKE_BUILD_TYPE=Release  # Release build
    cmake                             # armhf/arm64 environment

ASAN default on:

    cmake                             # X86 environment

However, if you manually set the options as shown below, you can set
the ASAN option values as desired in any environment.

    cmake -DENABLE_ASAN=OFF
    cmake -DENABLE_ASAN=ON

Signed-off-by: Inho Oh <inho.oh@sk.com>