SessionManager:Upgrade to handle Multi-Agent Participation
It upgrade to be possible to handle Participation
of Multi-Agent about session activation/deactivation.

For example, when the received directives are composed
of ASR and Display, both agents could activate or
deactivate session. Because the session is active
if any agent participate, for deactivating session,
all agents have to deactivate.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>