dialog:fix to handle previous dialog
If the previous dialog which is not handled exist,
it fix to handle that proper way.
 1) If playstack is stacked condition, it handle to complete.
 2) If playstack is not stacked condition, it handle to cancel.

Because, the processDirective of Capability execute important flow
about directive, it change to prevent overriding in derived class.

When the AudioPlayerAgent go to background focus,
if the player state is PREPARE, it cause skip to pause player,
so, media is playing continuously.
For preventing that situation, it add conditions about PREPARE, READY case.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>