TTS:fix handling sync state result
As some capability agents are operated with sync, there are cases
which need to to handle sync state result separately.

So, it fix not to handle the sync state result
if the condition is not satisfied in TTSAgent.

Also, it update to cancel previous directive
if the sync state result needs to suspend TTS play.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>