playsync:fix timer handling in PlaySyncManager test
In some build environment like docker, if the PlaySyncManager test
is executed, it's failed because the timing issue is occurred
when manipulating timer.

So, as adjusting timer flow and thread check mechanism,
it fix not to fail PlaySyncManager test.

Also, it reduce test elapsed time from 0.66 sec to 0.23 on average.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>