capability:change has_attachment variable name
It some case, it has to destroy directive by agent,
and the "has_attachment" variable is used for this purpose
in TTSAgent and AudioPlayerAgent.

But, in another case, even if it doesn't have attachment,
as it need to destroy directive byself,
the "has_attachment" variable name is not appropriate in this case.

So it change that variable name to destroy_directive_by_agent.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>