core: add ANY medium to DirectiveSequencer
The `BlockingMedium::ANY` type was added to the spec document.
This type is blocked by all other types(Audio/Visual/None/Any),
and all other types can be blocked.

e.g. When the sequencer receives a list of directives as below
for the dialog ID:

    [
      1:Audio/block,
      2:Any/block,
      3:None/non-block,
      4:Any/block,
      5:None/non-block
    ]

Directives are processed in the following order.

    Start-1 -> Complete-1
    -> Start-2 -> Complete-2
    *)-> Start-3 -> Complete-3
    -> Start-4 -> Complete-4
    -> Start-5 -> Complete-5

*) Even in the case of a non-block, if there is an ANY behind it,
   the ANY must be processed after the non-block is completed.

Signed-off-by: Inho Oh <inho.oh@sk.com>