core: add cancel by group to directive sequencer
Currently, the cancellation API provided by DirectiveSequcner cancels
the entire directive corresponding to the dialog ID, but an API that
cancels only specific directives is also required.

So, add the following API.

    bool cancel(const std::string& dialog_id,
                std::set<std::string> groups)

For groups, you can use a list of strings in the form of
`"{Namespace}.{Name}"` such as `"PhoneCall.MakeCall"`.

    cancel("dlg-1", { "PhoneCall.MakeCall", "PhoneCall.EndCall" });

Signed-off-by: Inho Oh <inho.oh@sk.com>