network: add on-demand network type
Currently, only the method of maintaining the connection was
supported, but due to a new requirement, a method of requesting
a connection only when necessary was added. Aka connectionless
or on-demand connection.

When a network connection request is received from a user, the
gateway list is received from the registry server and the network
status is changed to `READY`.

After receiving an event transmission request, the SDK connects to
the gateway server, transmits the event, and terminates the
connection after receiving all responses.

Signed-off-by: Inho Oh <inho.oh@sk.com>