TextAgent:Apply focus in sending TextInput event
When the ASR has foreground focus, if it try to send TextInput event,
the ASR is canceled and the focus is released.

In this situation, if background media exist, it has acquire foreground
focus even if not receiving response of TextInput event, so,
it cause abnormal behavior.

So, it update to relay focus in TextInput event, if the ASR has foreground
focus for maintaining focus until receiving response.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>