AudioPlayer:update condition to skip intermediate foreground focus
When sending StopCommandIssued and PauseCommandIssued event,
if the focus is mixed with ASR, before receiving related directive,
it could chance to get foreground focus intermediately,
so it play media in very short term.

For preventing this abnormal situation, it apply flag not to play media
even if it has foreground focus until receiving response of event.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>