CapabilityManager:Handle Network Disconnected Status
In some case, there needs to reset some variables in CapabilityAgent
when the network is disconnected.

So, it implements onStatusChanged callback in CapabilityManager
and send the network_disconnected command to each Agent in above case.

In currently, the AudioPlayerAgent needs that command,
it update to handle as resetting skip_intermediate_foreground_focus.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>