capability: add cancel option to stopRecognition
If the cancel option is true, all directives corresponding to the
current dialog are canceled among the directives to be received in
the future. It also cancels directives that have already been
received but are still pending.

However, directives that have already been received and are in the
processing state (active state) are not canceled and remain.
In this case, it must be stopped manually via the corresponding
agent's function as needed.

Signed-off-by: Inho Oh <inho.oh@sk.com>