network: add FAILED error status
When a connection attempt fails for the entire device-gateway
server list, there is no separate FAILED event and there is only
a DISCONNECT event, so it is difficult to determine the exact
status in the application.

To solve this problem, add a FAILED event, and when the application
receives this event, it checks the network status and tries to
connect again.

Signed-off-by: Inho Oh <inho.oh@sk.com>