base: add null check to plugin_find
If `-1` is returned in the `init` handler of a specific plug-in
during the initialization process of all plug-ins, the plug-in is
checked for removal and is removed from the plug-in list after
initialization.

At this time, when searching with the `nugu_plugin_find` API in
other plugins, the list item may be `null` in the case of a plugin
scheduled to be removed.

    Plugin A: return -1 in init()
    Plugin B: nugu_plugin_find('A') in init() -> crash

Signed-off-by: Inho Oh <webispy@gmail.com>