fix last attachment data for ASR
When recognizing the end of a voice, a separate end attachment packet
was transmitted without body data, but it was inefficient because it
had to send an unnecessary header.

In order to reduce unnecessary network transmission, it has been
modified to send the end mark in the header together with the last
data when recognizing the end of the voice.

Signed-off-by: Inho Oh <webispy@gmail.com>