Capability:Add API about capability handler in ICapabilityListener
It add the getter/setter API of capability handler in ICapabilityListener.
* void setCapabilityHandler(ICapabilityInterface* handler);
* ICapabilityInterface* getCapabilityHandler();

Because, API for capability handler is added in ICapabilityListener,
there is no need to use CapabilityListener anymore. So it's removed.

As the ICapabilityListener might be inherited with any capability listener
it change to be inherited virtually in each capability listener interface.

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>