upgrade clang-tidy-action to v1.2
It upgrade the clang-tidy-action to v1.2 activating on ubuntu 20.04.

It upgrade the clang-tidy-action to v1.2 activating on ubuntu 20.04.

It fix to use const/universal reference for copied value usage detection.
* performance-for-range-copy
* performance-unnecessary-value-param

It suppress the C-style variadic function usage because it seems fair use.
(It would upgrade clang-tidy-action to be able to handle this properly.)
* cert-dcl50-cpp

Signed-off-by: Hyungrok.Kim <hr97gdi@sk.com>