base: add log prefix to syslog
To make log analysis more convenient in the syslog system, it has
been modified to support prefix output such as thread-id and filename
in the same way as stderr and stdout.

However, unlike other log systems, prefix is output only when the
`NUGU_LOG_PREFIX` is set as an environment variable.

    export NUGU_LOG=syslog
    export NUGU_LOG_PREFIX=filename,line,tid

Signed-off-by: Inho Oh <inho.oh@sk.com>